README appropriate to obs_mosaic
Create 'exp' datatype from the /home/DLS preprocessed image

Applies the bpms from home/DLS for the observat and year as BAD
Applies the manual exclusion regions from home/DLS as SUSPECT
set sat bit for pixels > 20000
Remove the "exp" dataset and add the postISRCCD

Implement a MosaicMapper std_preprocessed to make the dataset consistent
with requirements of isr.
Handles the loading of the header information from the MEF file
created by the DLS pipeline.  In particular, the FILTER, OBSERVAT,
EXPTIME, and DATA-OBS are needed from hdu0 for the LSST pipeline.

Create a custom visitInfo with this information.

Custom Isr task for Mosaic2.  mosaicPreprocessedIsr adds the bpm from the
correct observatory and the .reg files from home/DLS

Removed some of the datatypes not needed by Mosaic.  Many were only
used for Decam.